requirements.txt
PyAutoGUI~=0.9.52
Pillow~=8.1.2
numpy~=1.20.1